Send teardown time in FlowLog messages.
If a Flow Log message is exported for a flow, ensure that
flow teardown time is also sent in subsequent exports. If we
encounter a flow-sample with teardown time for a flow which was
never exported, then subject it sampling just like other flows.

Also treat -1 value for flow-export-rate as disabling of sampling.

Closes-Bug: #1578982

+++ src/vnsw/agent/test-xml/test_xml_oper.cc
@@ -885,7 +885,7 @@ bool AgentUtXmlGlobalVrouterValidate::Validate() {
         return false;
 
     if (flow_export_rate_ != -1) {
-        uint32_t rate = (uint32_t)(flow_export_rate_);
+        int32_t rate = flow_export_rate_;
         if (vr->flow_export_rate() != rate)
             return false;
     }
